cache challenges, flip and connect with profile
