+++ package.json
@@ -13,30 +13,31 @@
     "@expo/vector-icons": "^10.0.0",
     "expo": "^35.0.0",
     "expo-cli": "^3.8.0",
+    "expo-constants": "^7.0.1",
     "expo-facebook": "~7.0.0",
+    "expo-font": "~7.0.0",
     "expo-linear-gradient": "~7.0.0",
     "expo-svg-uri": "^1.1.0",
     "firebase": "^7.4.0",
-    "expo-font": "~7.0.0",
     "native-base": "^2.13.8",
     "prop-types": "^15.7.2",
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-    "react-native-svg": "9.9.5",
-    "react-native-web": "^0.11.7",
+    "react-native-card-flip": "^1.0.6",
     "react-native-gesture-handler": "~1.3.0",
     "react-native-reanimated": "~1.2.0",
     "react-native-screens": "~1.0.0-alpha.23",
     "react-native-snap-carousel": "^3.8.4",
+    "react-native-svg": "9.9.5",
+    "react-native-web": "^0.11.7",
     "react-navigation": "^4.0.10",
     "react-navigation-stack": "^1.10.3",
     "react-navigation-tabs": "^2.5.6"
   },
   "devDependencies": {
-    "babel-preset-expo": "^7.1.0",
-    "react-native-svg-transformer": "^0.14.1",
     "babel-eslint": "^9.0.0",
+    "babel-preset-expo": "^7.1.0",
     "eslint": "^5.14.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^4.1.0",
@@ -47,7 +48,8 @@
     "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.12.4",
     "eslint-plugin-react-hooks": "^1.3.0",
-    "prettier": "^1.16.4"
+    "prettier": "^1.16.4",
+    "react-native-svg-transformer": "^0.14.1"
   },
   "private": true,
   "eslintConfig": {
